Re #600
Adding a reset tracker button to the device set selector dialog. It is only visible when the connected tracker says that it is resettable.

Currently only implemented for the Phidget accelerometer based tracking system.

Can be implemented for any system by overriding the IsResettable() and Reset() functions of any vtkPlusDevice