re #618: Two changes:
(1) I changed from using amplitude normalization to std normalization. This change was made because it seems that normalization by std is more robust than by amplitude (based on our collected US sequences). As a result normalization by std produces lower variances in time-offset between trials (for fixed imaging parameters; again, based on our collected US sequences).

(2) I changed the max time-offset from 2 s to 0.5 s. Note that the 0.5 s max-time offset is desirable because (a) the time-offsets are all under 0.5 s (and thus within the +/-0.5 s range created by a max-time offset of 0.5 s) and (b) the +/-0.5 s range allows for a period of probe motion that is as low as 1 s. On the other hand, with a 2 sec max time-offset, we must assume that the period of probe motion is greater than 4 s(which it often isn't). Note that if the period of probe motion is less than 2 times the max time-offset and the signal is (perfectly/ nearly) periodic, there is a risk that the sliding signal is shifted over by a period during the alignment process (i.e. the computed time-offset will be: true time-offset + 1 period of probe motion). I changed the max time-offset for 2 s to 0.5 s both in the TemporalCalibrationAlgoTest.cxx file and as the default in the TemporalCalibrationAlgo.cxx file.