re #840: Fixed Windows SDK detection
The FindWindowsSDK.cmake in Utilities caused build error with VS2008, as it returned the wrong SDK version (6.1). The issue is this find macro can only return the CURRENT version or a specific version. So, just keep using the FindWindowsSDK.cmake that is already in PlusBuild and works better (can return the LATEST SDK version).

Removed SDK detection error check, because it did not work (did not print an error when the wrong SDK was used, the build failed at the unspecified constant in an FormatReader.cpp). The problem is that when multiple SDKs are detected it's not enough to check if 7.1 is detected but it should be checked what is actually used. Moreover, any SDK version >7.1 should be accepted, too, so this check actually prevents some correct configuration from building. Implementing a correct solution would be difficult, so removed this check and just added a comment in FormatReader.cpp where the error occurs that SDK 7.1 or more recent has to be installed (it's in the build instructions, too).

Moved required MF link libraries to the MfVideoCaptureLibrary.