re #696: Fixed issue with ICCapturing device not taking into account PortUsImageOrientation.
vtkICCapturingSource is a singleton and therefore the settings are only read the first time after application start. When we disconnect/connect the old channels are still there, so if something changed in the config file it has no effect. The fix was to clear the existing channels and tools at the beginning of configuration reading.