re #762: Fixed race condition at logger initialization
Problem was that AccurateTimer accessed the logger first, but while performing initialization, the application's main thread accessed the logger. The logger creation mutex was different from the logger access mutex and therefore the logger was used before fully initialized. The only side effect that was observed is that setting verbose level on the command line was ignored (because it was set from the main thread but later it was overwritten by the logger initialization process).

Solution: lock the logger access mutex as well until logger initialization is fully completed. The access mutex is locked even before setting the logger instance variable, so there is no time window at all for any thread to get access to the logger before it is initialized.