enable usage of openigtlink install-type files
This commit modifies the include statements, so it is compatible with the installed version of openigtlink (as compared to only the built version). This is due to the fact, that OpenIGTLink installs header files to different locations in the install "distribution", specifically, the igtlutil subfolder is just installed into the base include folder.