This update makes it so the program can test vtkPlusTransverseProcessEnhancer.cxx from the command line.
This program can be given an input file, a config file, parameters, and an output file. This program applies the parameters to vtkPlusTransverseProcessEnhancer, which then applies the changes to the input file, and saves the changes to the output file. All parameters are saved to the config file.

Added tests to the following vtkPlusTransverseProcessEnhancer methods in the following order:
-WriteConfiguration
-ReadConfiguration
-ProcessFrame