Re #1108 Start of scoliosis-transverse-process-enhancer branch. Recommitting all changes under new branch
Built file vtkPlusTransverseProcessEnhancerTest.cxx so that it may be used to test vtkPlusTransverseProcessEnhancer.cxx with greater ease.

Tests can be done via the command line by giving a config file, an input file, and an output file. Optional additional files for additional output data include a shadow image file, a lines image file, a processed lines image file, and an intermediate image file. Note, the previously mentioned optional output files may not work due to them not functioning well enough in vtkPlusTransverseProcessEnhancer.cxx

vtkPlusTransverseProcessEnhancer.cxx has also been changed so that it may better write to the config file it is given in the method WriteConfiguration()
vtkPlusTransverseProcessEnhancer.h has also been changed to allow for this to occur