re #1108 Fully connected the pipeline. This update doesn't fully work.
Added IslandRemover, ImageEroder, and ImageDialator steps to the pipeline, though they currently only update to a blank image.

Also moved the steps where preceding bone images, and bone images that are off camera are removed. They now happen after the ImageDialator step so as to not interfere with the pipeline.