re #1108 Fixed issues with the pipeline, and added a check that compares Bone areas to the areas beside them.
In the last commit, the pipeline was fragmented and did not work properly. This issue has since been fixed.
There is also a new function that checks the amount of shadow found in a bone area and compares it to the amount of bone in the areas around it. If the areas surrounding the bone have more shadow, then the area is no longer considered to be a potential bone area.
Also changed the search algorithm so it is more accurate, but recursive.