re #1108 Bone area search is now based on slope. Added a second island removal step.
The location of bone areas is now determined when pre-bone images are removed, and saved in a class variable for faster, easier access in future. The start and end of bone areas is now determined based on their relative slope to the pixels around them, rather then if they are connected.
Note: the previous way to find bone areas is still in the code, it is just not accessible.

Another island removal step was added after the erosion step to remove smaller islands.

The step for comparing a bone area to the areas next to it was made more flexible, to accommodate large bone areas close it the edge of the screen