re #246: Disable PlusServerOpenIGTLinkCommandsTest on Linux
The test hangs on Linux waiting for mutex locks in vtkPlusOpenIGTLinkClient::SendCommand() and vtkPlusOpenIGTLinkClient::SendMessage().
This only happens when the PlusServer process is launched from PlusServerRemoteControl, but not when it's started separately.
For now, the conclusion is that we will disable the test on Linux, and potentially try to fix this at a later time.