ENH: Use ParameterInfo structure in US imaging parameters instead of two maps
Added Changed flag to parameter info. It will be used to indicate that a parameter is changed, but not set to the US device yet.
Fixed recognition of US parameter names: comparison is now case-insensitive, so finds the correct entry in the map (used to create new entry if difference was only in case)