fix #549: BK ultrasound device compile error for installed VTK
Private png headers are not copied during VTK install step, however the BK device requires the private png implementation details.

This adds a link to the private headers in the source using vtkPNGPrivate.h.in, which is configured to include the absolute path to pngpriv.h. (Configuring requires that VTK_SRC_DIR is defined).