ENH: Return immediately after ARFIPush
It was hardcoded to 8 seconds so that upon exiting the method the frame would be available in the frame list. This time was specific to the previously hardcoded 30 pushes method. Push configuration string can be customized now so a time to add to the frame list is dynamic.

If trying to get the frame from the frame list, the video source should be checked to see if it is empty before attempting to get the frame.