ENH: Update expected ARFI timeblock size
When there was ARFI support for X4 and X8BF, the timeblock size was the same for both where the X4 needed to use the full size and X8 used half the size and filled in the other half with zeros.  Since X4BF ARFI support was recently removed, the data now has a timeblock size that is half of what is was before.