ENH: Changes to Atracsys wrapper (#1108)
Clarification that the maximum number of events is actually the maximum number of additional events on top of the default allocation of 20.

Removed the event synthTempv1 which is obsolete.

The number of fiducials is now indicated in the "Marker_Info" framefield.

Option names that are not included in the dictionary can now be parsed and applied. Simply replace the spaces by underscores to set any option compatible with the SDK/device.